+++ lib/google_maps_utils.dart
@@ -2,6 +2,5 @@
 library google_maps_utils;
 
 export 'poly_utils.dart';
-export 'lat_lng.dart';
 export 'math_utils.dart';
 export 'spherical_utils.dart';